change!: Move `object::tree::diff::change::DiffPlatform` to `object::blob::diff::Platform`.
